requirements.txt
openpyxl==3.0.10
pandas==2.0.3
pandas==1.4.3
Pillow==9.2.0
Pillow==10.0.0
plotly==4.14.3
streamlit==1.12.0
streamlit_option_menu==0.3.6
